+++ screenobjects/myEventsList.screen.js
@@ -1,12 +1,13 @@
-export default class EventList {
+import Basics from './basic.screen'
+export default class MyEventsList extends Basics {
     get_what_to_seek () {
         switch (browser.capabilities.platformName) {
             case "Android":
             case "android":
-                return '//android.widget.TextView[@text="Event list"]'
+                return '//android.widget.TextView[@text="Your events"]'
             case "iOS":
             case "ios":
-                return '//XCUIElementTypeStaticText[@name="Listado de eventos"]'
+                return '//XCUIElementTypeStaticText[@name="DashboardView.myEvents" and @label="Tus eventos"]'
         }
     }
     get_where_tap_on(where){
@@ -20,6 +21,9 @@ export default class EventList {
                 case "ios":
                     return '//XCUIElementTypeButton[1]'
             }
+           
+            default:
+                return super.get_where_tap_on(where)
         }
     }
 }
